Commit# 7 - JwtExtension change & route changes   * Moved the JwtExtension from Authentication.API to a separate library under utilities folder   * Added assign task & update allocation navigation to Ocelot config   * Fixed response of assign task and Add member endpoints   * Updated Endpoint route for Update Allocation
