+++ ApiGateways/OcelotApiGateway/OcelotApiGateway.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Authentication\Authentication.API\Authentication.API.csproj" />
+    <ProjectReference Include="..\..\Utilities\JwtExtensions\JwtExtensions.csproj" />
   </ItemGroup>
 
 </Project>
